opsInFlight values should be a Set; tests
